[refactor] remove c++ from AurorOS
