+++ include/apps/calc.h
@@ -1,0 +1,1 @@
+void calculator_main();
