--- include/apps/calc.hpp
@@ -1,1 +1,0 @@
-extern "C" void calculator_main();
