+++ include/apps/tinypad.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <screen.h>
+#include <types.h>
+#include <input.h>
+
+int tinypad_main(uint8_t color, uint8_t barcolor);
