--- include/apps/tinypad.hpp
@@ -1,19 +1,0 @@
-/**
- * -------------------------------------------------------------------------
- *                                   AurorOS
- * (c) 2022-2025 Interpuce
- * 
- * You should receive AurorOS license with this source code. If not - check:
- *  https://github.com/Interpuce/AurorOS/blob/main/LICENSE.md
- * -------------------------------------------------------------------------
- */
-
-#pragma once
-
-extern "C" {
-    #include <screen.h>
-    #include <types.h>
-    #include <input.h>
-}
-
-int tinypad_main(uint8_t color, uint8_t barcolor);
