--- src/apps/shell/commands/commands.hpp
@@ -1,19 +1,0 @@
-#pragma once
-
-#include <types.h>
-#include <screen.h>
-#include <fs/filesystem.h>
-
-namespace ShellCommands {
-    extern void cowsay(char message[1024]);
-    extern void map();
-    extern void eclair(const char *trt);
-    extern void help();
-    extern void cat(fs_node* current_dir, char* where, uint64_t current_user_id);
-    extern void cd(fs_node** current_dir, char* where, uint64_t current_user);
-    extern void ls(fs_node* dir);
-    extern const char* pwd(fs_node* current_dir);
-    extern void mkdir(fs_node* current_dir, const char* where, uint64_t owner);
-    extern void rm(fs_node* current_dir, const char* path);
-    extern void chmod(fs_node* current_dir, uint64_t current_user_id, char* perm_arg, char* target_path);
-}
